fix: rename hook
