+++ src/app/signup/terms-of-service/hooks/useTermsOfServiceFormViewModel.ts
